backend/requirements.txt
# Web framework
fastapi==0.129.0
uvicorn[standard]==0.34.0

# LangChain / LangGraph / Gemini
langchain-core==0.3.28
langchain-google-genai==2.1.3
langgraph==0.2.60
google-generativeai>=0.8.0

# Database — PostgreSQL
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Redis — distributed session state
redis==5.2.1

# Rate limiting
slowapi==0.1.9

# Config
python-dotenv==1.0.1
pydantic==2.10.4
